+++ src/elements/Header/Header.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import removeExtraSpaces from '../../helpers/removeExtraSpaces';
 import './Header.css';
 
 const getSize = (props) => {
@@ -6,18 +7,24 @@ const getSize = (props) => {
   if (props.md) return 'md';
   if (props.lg) return 'lg';
   return 'sm'; // fallback default size
+};
+
+const defaultProps = {
+  sm: false,
+  md: false,
+  lg: false,
+  className: '',
 }
 
-export default function Header(props) {
+function Header(props) {
   const { className = '', children } = props;
   const customClassName = `
     Header
     ${className}
     ${getSize(props)}
   `;
-  return (
-    <h1 className={customClassName}>
-      {children}
-    </h1>
-  )
+  return <h1 className={removeExtraSpaces(customClassName)}>{children}</h1>;
 }
+
+Header.defaultProps = defaultProps;
+export default Header;