+++ CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add `ProgressBar` element
+- Add `Icon` element
+- Add `Text` element
+- Add `ProfilePicture` element
+- Add `Menu` component
+- Add `SocialItem` element
+- Add `Button` element - primary, secondary and default styles
 - Add `Header` element
 
 ## [0.0.1] - 2021-02-20
@@ -20,5 +27,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add README.md file
 - Remove unnecessary boilerplate files
 
-[Unreleased]: https://github.com/npranto/portfolio-marlon/compare/v0.0.1...npranto:master
+[unreleased]: https://github.com/npranto/portfolio-marlon/compare/v0.0.1...npranto:main
 [0.0.1]: https://github.com/npranto/portfolio-marlon/tree/v0.0.1